+++ ts-client/src/test/sdk_token2022.test.ts
@@ -30,6 +30,7 @@ import {
   VersionedTransaction,
 } from "@solana/web3.js";
 import BN from "bn.js";
+import { randomInt } from "crypto";
 import Decimal from "decimal.js";
 import fs from "fs";
 import { DLMM } from "../dlmm";
@@ -37,7 +38,6 @@ import {
   BASIS_POINT_MAX,
   ConcreteFunctionType,
   DEFAULT_BIN_PER_POSITION,
-  FunctionType,
   LBCLMM_PROGRAM_IDS,
   POSITION_MAX_LENGTH,
 } from "../dlmm/constants";
@@ -60,6 +60,7 @@ import {
 } from "../dlmm/helpers/token_2022";
 import {
   ActivationType,
+  BinAndAmount,
   MEMO_PROGRAM_ID,
   ResizeSide,
   StrategyType,
@@ -76,7 +77,6 @@ import {
   OperatorPermission,
   sendTransactionAndConfirm,
 } from "./helper";
-import { randomInt } from "crypto";
 
 const keypairBuffer = fs.readFileSync(
   "../keys/localnet/admin-bossj3JvwiNK7pvjr149DqdtJxf2gdygbcmEPTkb2F1.json",
@@ -2863,6 +2863,164 @@ describe("SDK token2022 test", () => {
         // Due to transfer fee
         expect(consumedTokenX.gte(totalXAmount)).toBeTruthy();
         expect(consumedTokenY.lte(totalYAmount)).toBeTruthy();
+      });
+
+      it("Add liquidity by weight2 chunk", async () => {
+        const totalXAmount = new BN(5).mul(new BN(10 ** btcDecimal));
+        const totalYAmount = new BN(5).mul(new BN(10 ** solDecimal));
+
+        const dlmm = await DLMM.create(connection, pairKey, opt);
+
+        const positionKp = Keypair.generate();
+
+        const minBinId = dlmm.lbPair.activeId - 200;
+        const maxBinId = dlmm.lbPair.activeId + 200;
+
+        let tx = await dlmm.createExtendedEmptyPosition(
+          minBinId,
+          maxBinId,
+          positionKp.publicKey,
+          adminKeypair.publicKey,
+        );
+
+        await sendAndConfirmTransaction(connection, tx, [
+          adminKeypair,
+          positionKp,
+        ]);
+
+        const xAmountBps = Math.floor(
+          10_000 / (maxBinId - dlmm.lbPair.activeId + 1),
+        );
+        const yAmountBps = Math.floor(
+          10_000 / (dlmm.lbPair.activeId - minBinId + 1),
+        );
+
+        const xYAmountDistribution: BinAndAmount[] = [];
+
+        let totalXBps = new BN(0);
+        let totalYBps = new BN(0);
+
+        for (let binId = minBinId; binId <= maxBinId; binId++) {
+          if (binId < dlmm.lbPair.activeId) {
+            xYAmountDistribution.push({
+              binId,
+              xAmountBpsOfTotal: new BN(0),
+              yAmountBpsOfTotal: new BN(yAmountBps),
+            });
+
+            totalYBps = totalYBps.add(new BN(yAmountBps));
+          } else if (binId > dlmm.lbPair.activeId) {
+            xYAmountDistribution.push({
+              binId,
+              xAmountBpsOfTotal: new BN(xAmountBps),
+              yAmountBpsOfTotal: new BN(0),
+            });
+
+            totalXBps = totalXBps.add(new BN(xAmountBps));
+          } else {
+            xYAmountDistribution.push({
+              binId,
+              xAmountBpsOfTotal: new BN(xAmountBps),
+              yAmountBpsOfTotal: new BN(yAmountBps),
+            });
+
+            totalXBps = totalXBps.add(new BN(xAmountBps));
+            totalYBps = totalYBps.add(new BN(yAmountBps));
+          }
+        }
+
+        const xBpsLoss = new BN(10_000).sub(totalXBps);
+        const yBpsLoss = new BN(10_000).sub(totalYBps);
+
+        xYAmountDistribution[0].yAmountBpsOfTotal =
+          xYAmountDistribution[0].yAmountBpsOfTotal.add(yBpsLoss);
+        xYAmountDistribution[
+          xYAmountDistribution.length - 1
+        ].xAmountBpsOfTotal =
+          xYAmountDistribution[
+            xYAmountDistribution.length - 1
+          ].xAmountBpsOfTotal.add(xBpsLoss);
+
+        let txs = await dlmm.addLiquidityByWeight2(
+          {
+            positionPubKey: positionKp.publicKey,
+            totalXAmount,
+            totalYAmount,
+            xYAmountDistribution,
+            user: adminKeypair.publicKey,
+            slippage: 0,
+          },
+          // 50 bin per chunk
+          50,
+        );
+
+        const userXAta = getAssociatedTokenAddressSync(
+          dlmm.lbPair.tokenXMint,
+          adminKeypair.publicKey,
+          true,
+          dlmm.tokenX.owner,
+        );
+
+        const userYAta = getAssociatedTokenAddressSync(
+          dlmm.lbPair.tokenYMint,
+          adminKeypair.publicKey,
+          true,
+          dlmm.tokenY.owner,
+        );
+
+        const [beforeUserXAccount, beforeUserYAccount] =
+          await connection.getMultipleAccountsInfo([userXAta, userYAta]);
+
+        const txSigs = await Promise.all(
+          txs.map((tx) => {
+            return sendAndConfirmTransaction(connection, tx, [adminKeypair]);
+          }),
+        );
+
+        const [afterUserXAccount, afterUserYAccount] =
+          await connection.getMultipleAccountsInfo([userXAta, userYAta]);
+
+        const beforeUserXTokenAccount = unpackAccount(
+          userXAta,
+          beforeUserXAccount,
+          dlmm.tokenX.owner,
+        );
+
+        const beforeUserYTokenAccount = unpackAccount(
+          userYAta,
+          beforeUserYAccount,
+          dlmm.tokenY.owner,
+        );
+
+        const afterUserXTokenAccount = unpackAccount(
+          userXAta,
+          afterUserXAccount,
+          dlmm.tokenX.owner,
+        );
+
+        const afterUserYTokenAccount = unpackAccount(
+          userYAta,
+          afterUserYAccount,
+          dlmm.tokenY.owner,
+        );
+
+        const depositedXAmount =
+          beforeUserXTokenAccount.amount - afterUserXTokenAccount.amount;
+        const depositedYAmount =
+          beforeUserYTokenAccount.amount - afterUserYTokenAccount.amount;
+
+        const xRatio = new Decimal(depositedXAmount.toString()).div(
+          new Decimal(totalXAmount.toString()),
+        );
+
+        const yRatio = new Decimal(depositedYAmount.toString()).div(
+          new Decimal(totalYAmount.toString()),
+        );
+
+        // Due to SDK bps to weight conversion + on chain precision loss
+        // When running single test iw will be 0.999 ... not sure which part causes the flakiness
+        expect(xRatio.toNumber()).toBeGreaterThan(0.88);
+        expect(yRatio.toNumber()).toBeGreaterThan(0.88);
       });
     });
   });