+++ CHANGELOG.md
@@ -52,6 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added function `enablePositionPermissionlessClaimFee`. This allow non position owner to invoke claim fee instruction for other's position.
 - Added function `getLimitOrderByUserAndLbPair` to get all limit orders of a user for a given pair.
 - Added function `getLimitOrder` to get limit order data by limit order pubkey.
+- Added function `addLiquidityByWeight2` which support token 2022 add liquidity by weight.
 
 ### Removed